ref: divide queries file into different files
